requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pymongo==4.12.1
python-jose==3.4.0
passlib==1.7.4
python-multipart==0.0.20
bcrypt==4.3.0
pydantic==2.11.4
